Split apart to support Union types
Since the types contained within a union cannot be collections

# Conflicts:
#	core/src/main/groovy/org/grails/gorm/graphql/entity/dsl/helpers/ComplexTyped.groovy
